shared/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alphamachine-core"
version = "0.1.0"
description = "Core utilities for Alpha Machine"
authors = [
  { name = "Jonathan", email = "jonathan@example.com" },
]
license = { file = "LICENSE" }
dependencies = [
    "python-dotenv",
    "pyyaml"
]

[project.urls]
"Homepage" = "https://github.com/your-repo/alpha-machine"
"Bug Tracker" = "https://github.com/your-repo/alpha-machine/issues"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."] 